write stream to a fifo
create a fifo at the tmppath and write the stream to this file

the writer is blocking, so we write the first segment to the fifo. This process blocks until the fifo is opened readable. After this is   done, the next segment is the oldest one in the queue. The max length of the queue is 10 segments. each segment has a length of 6 seconds(static!)